Updated plotters to remove basemap dependencies
All plotters for channels, geocolor and webmet were updated to remove
basemap dependencies and use cartopy. This is done in most cases by
switching to plt.axis(projection=ccrs.PlateCarree()) and with a custom
extent using cartopy conventions extente = [west, east, south, north].

In the case of each channel, the colobar was defined ad-hoc to look like
the current colorbar in the website https://wrf.ohmc.ar/goes/C08/.
However a better solution might be found.

All places with changes were marked with #TODO and missing files
were indentified to be asked for.